requirements.txt
python-telegram-bot
yadisk
requests
asyncio